+++ app/controllers/api/v2/ExtAuthController.js
@@ -78,7 +78,7 @@ export const authFinish = compose([
       const profileData = {
         provider:   provName,
         externalId: state.profile.id,
-        title:      state.profile.fullName,
+        title:      state.profile.name,
       };
 
       // Connect external profile to FreeFeed account
@@ -93,7 +93,7 @@ export const authFinish = compose([
 
         if (profileUser && profileUser.id !== currentUser.id) {
           throw new ForbiddenException(
-            `The '${state.profile.fullName}' profile on ${authProvider.title} is already ` +
+            `The '${state.profile.name}' profile on ${authProvider.title} is already ` +
           `associated with another FreeFeed account: @${profileUser.username}`
           );
         }
@@ -106,7 +106,7 @@ export const authFinish = compose([
       if (state.params.mode === MODE_SIGN_IN) {
         const profile = {
           provider:   provName,
-          name:       state.profile.fullName,
+          name:       state.profile.name,
           email:      state.profile.email,
           pictureURL: state.profile.pictureURL,
         }
@@ -163,8 +163,6 @@ export const authFinish = compose([
       }
 
       throw err;
-    } finally {
-      await authProvider.done(ctx.request.body);
     }
   },
 ]);