+++ app/controllers/api/v2/ServerInfoController.js
@@ -1,11 +1,11 @@
-import config from 'config'
-
 import { version } from '../../../../package.json';
 import { ServerInfo } from '../../../models';
+import { allExternalProviders } from '../../../support/ExtAuth';
 
 
 export async function serverInfo(ctx) {
-  const externalAuthProviders = Object.keys(config.externalAuthProviders || {});
+  const externalAuthProviders = allExternalProviders
+    .map(({ id, title }) => ({ id, title }));
   const registrationOpen = await ServerInfo.isRegistrationOpen();
   ctx.body = {
     version,