+++ test/functional/server-info.js
@@ -1,10 +1,10 @@
 /* eslint-env node, mocha */
 import fetch from 'node-fetch';
 import expect from 'unexpected';
-import config from 'config';
 
 import { getSingleton } from '../../app/app';
 import { version as serverVersion } from '../../package.json';
+import { allExternalProviders } from '../../app/support/ExtAuth';
 
 import { serverInfoResponse } from './schemaV2-helper';
 
@@ -31,7 +31,8 @@ describe('/v2/server-info', () => {
 
   it(`should return the externalAuthProviders`, async () => {
     const resp = await fetch(`${host}/v2/server-info`).then((r) => r.json());
-    const externalAuthProviders = Object.keys(config.externalAuthProviders || {});
-    expect(resp.externalAuthProviders.sort(), 'to equal', externalAuthProviders.sort());
+    const externalAuthProviders = allExternalProviders
+      .map(({ id, title }) => ({ id, title }));
+    expect(resp.externalAuthProviders, 'to equal', externalAuthProviders);
   });
 });